web/homework_9/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
lxml==5.2.1
mongoengine==0.28.2
pymongo==4.7.1
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
